perf-analysis: deduplicate get_static_symbols and subs_till_fixed_point
The combined PR introduced two copies of get_static_symbols and
subs_till_fixed_point: an older one in helpers.py (used by work_depth and
operational_intensity) with a hard-coded type-name list and raw sympy, and an
improved one in total_volume.py deriving the type names from dace.dtypes and
using dace.symbolic. Keep a single improved implementation in helpers.py (the
shared module) and import it from total_volume.

get_static_symbols keeps returning string-keyed mappings, which both
substitution (expr.subs) and by-name indexing (mapping[loop_var] in
operational_intensity) rely on, so all three analyses now share the better
type-cast handling with no change to their contracts.
