+++ dace/sdfg/performance_evaluation/total_volume.py
@@ -20,7 +20,6 @@ fully naive per-scalar-access count. """
 
 import argparse
 import os
-import re
 from collections import deque
 from copy import deepcopy
 from typing import Dict, List, Optional, Tuple
@@ -32,6 +31,7 @@ from dace.data import View
 from dace.dtypes import StorageType
 from dace.sdfg import infer_types, nodes as nd
 from dace.sdfg.graph import MultiConnectorEdge
+from dace.sdfg.performance_evaluation.helpers import get_static_symbols
 from dace.sdfg.state import (AbstractControlFlowRegion, BreakBlock, ConditionalBlock, ContinueBlock, LoopRegion,
                              ReturnBlock)
 from dace.symbolic import pystr_to_symbolic, symbol
@@ -40,24 +40,6 @@ from dace.transformation.passes.analysis import loop_analysis
 
 RegionVolumeMap = Dict[AbstractControlFlowRegion, Tuple[sp.Expr, sp.Expr]]
 RangeVarStack = List[Tuple[str, Tuple[sp.Expr, sp.Expr, sp.Expr]]]
-
-
-def subs_till_fixed_point(expr: sp.Expr, symbol_map: Dict[sp.Expr, sp.Expr]) -> sp.Expr:
-    """
-    Apply a symbol mapping to a symbolic expression repeatedly until a fixed point is reached.
-
-    Requires that the symbol mapping has no cyclic dependencies, otherwise it would not converge.
-
-    :param expr: The expression to substitute into.
-    :param symbol_map: Mapping from symbols to their replacement expressions.
-    :return: The expression after substituting to a fixed point.
-    """
-    prev = None
-    curr = expr
-    while prev != curr:
-        prev = curr
-        curr = curr.subs(symbol_map)
-    return curr
 
 
 def safe_summation(summand: sp.Expr, var: sp.Symbol, lower: sp.Expr, upper: sp.Expr) -> sp.Expr:
@@ -147,88 +129,6 @@ def resolve_minmax_over_range(expr: sp.Expr, var: sp.Symbol, lower: sp.Expr, upp
                 replacements[node] = original_operand
                 break
     return expr.subs(replacements) if replacements else expr
-
-
-def get_static_symbols(sdfg: SDFG) -> Dict[sp.Symbol, sp.Expr]:
-    """
-    Find the symbols that are assigned at exactly one point in the SDFG (i.e., statically known).
-
-    A symbol is static if it is written by a single length-1 access (from a tasklet performing one
-    assignment, or by a single copy from another access node). Symbols written in more than one
-    place are excluded.
-
-    :param sdfg: The SDFG for which to find static symbols and their assignments.
-    :return: Mapping from each static symbol to its defining expression (resolved to a fixed point).
-    """
-    # Strip type-cast prefixes (e.g. ``dace.float64``, ``int``) from a tasklet RHS so the cast does
-    # not interfere with the symbolic parse below. The DaCe type names are derived from
-    # ``dace.dtypes`` (rather than hard-coded), and matched longest-first so ``dace.float64`` wins
-    # over ``float``.
-    cast_names = {'int', 'float', 'complex', 'bool'}
-    cast_names |= {f'dace.{name}' for name in dir(dtypes) if isinstance(getattr(dtypes, name), dtypes.typeclass)}
-    type_regex = re.compile("|".join(re.escape(name) for name in sorted(cast_names, key=len, reverse=True)))
-
-    static_symbol_mapping: Dict[sp.Symbol, sp.Expr] = {symbol(a): symbol(a) for a in sdfg.arg_names}
-    non_static_symbols = set()
-    for node, containing_state in sdfg.all_nodes_recursive():
-        if not isinstance(node, nd.AccessNode):
-            continue
-        if containing_state.in_degree(node) != 1:
-            continue
-        edge = containing_state.in_edges(node)[0]
-        source = edge.src
-        if edge.data.volume != 1:
-            continue
-
-        if isinstance(source, nd.Tasklet):
-            tasklet = source
-            in_map = {}
-            out_map = {}
-            # Incoming edges: symbols feeding the tasklet.
-            for e in containing_state.in_edges(tasklet):
-                if not isinstance(e.src, nd.AccessNode):
-                    continue
-                in_map[e.dst_conn] = str(e.src.data)
-            # Outgoing edges: symbols written by the tasklet (expected to be a single edge).
-            for e in containing_state.out_edges(tasklet):
-                if not isinstance(e.dst, nd.AccessNode):
-                    continue
-                out_map[e.src_conn] = str(e.dst.data)
-
-            in_map = {symbol(k): symbol(v) for k, v in in_map.items()}
-            out_map = {symbol(k): symbol(v) for k, v in out_map.items()}
-            code = tasklet.code.as_string.strip()
-            # Expect a single assignment.
-            lines = [l.strip() for l in code.splitlines() if l.strip()]
-            if len(lines) > 1:
-                non_static_symbols.add(node.data)
-                continue
-            lhs, rhs = lines[0].split('=', 1)
-            lhs = lhs.strip()
-            rhs = type_regex.sub("", rhs.strip())
-            lhs_sympy = pystr_to_symbolic(lhs).subs(out_map)
-
-            if lhs_sympy not in static_symbol_mapping.keys():
-                try:
-                    static_symbol_mapping[lhs_sympy] = pystr_to_symbolic(rhs).subs(in_map)
-                except Exception:
-                    non_static_symbols.add(lhs_sympy)
-            else:
-                non_static_symbols.add(lhs_sympy)
-
-        elif isinstance(source, nd.AccessNode):
-            data_sym = symbol(source.data)
-            if data_sym not in static_symbol_mapping.keys():
-                static_symbol_mapping[data_sym] = symbol(node.data)
-            else:
-                non_static_symbols.add(data_sym)
-
-    static_symbol_mapping = {k: v for (k, v) in static_symbol_mapping.items() if k not in non_static_symbols}
-    static_symbol_mapping = {
-        k: subs_till_fixed_point(v, static_symbol_mapping)
-        for k, v in static_symbol_mapping.items()
-    }
-    return static_symbol_mapping
 
 
 def calculate_edge_volume(state: SDFGState, edge: MultiConnectorEdge) -> sp.Expr:
